fix: address PR review — regex collapse, move tests, fix comments
- Collapse dash replacement to single regex pass [^a-z0-9._-]+ → "-"
  (O(n) linear instead of O(n²) while loop)
- Move TestRootComponentPURL* + TestSanitizeNameForPurl from
  test_enrichment_module.py to test_augmentation_module.py
- Fix misleading terraform comment (leading dot preserved, substring match)
- Fix ambiguous "underscores preserved... or not?" comment
